+++ integration/src/test/kotlin/kotlinx/benchmark/integration/TransitiveDependenciesResolutionTest.kt
@@ -24,4 +24,9 @@ class TransitiveDependenciesResolutionTest : GradleTest() {
     fun wasmJs() {
         verifyFor("wasmJs")
     }
+
+    @Test
+    fun wasmWasi() {
+        verifyFor("wasmWasi")
+    }
 }