+++ runtime/commonMain/src/kotlinx/benchmark/Utils.kt
@@ -3,5 +3,3 @@ package kotlinx.benchmark
 internal expect fun String.readFile(): String
 
 internal expect fun String.writeFile(text: String)
-
-internal expect inline fun measureNanoseconds(block: () -> Unit): Long