+++ runtime/jsMain/src/kotlinx/benchmark/NodeJsEngineSupport.kt
@@ -10,9 +10,9 @@ private external object NodeFileSystem {
     fun readFileSync(path: String, options: String): String
 }
 
-internal object NodeJsEngineSupport : JsEngineSupport() {
-    override fun writeFile(path: String, text: String) =
-        NodeFileSystem.writeFileSync(path, text)
+private object NodeJsEngineSupport : BenchmarkEngineSupport() {
+    override fun writeFile(path: String, content: String) =
+        NodeFileSystem.writeFileSync(path, content)
 
     override fun readFile(path: String): String =
         NodeFileSystem.readFileSync(path, "utf8")
@@ -21,15 +21,30 @@ internal object NodeJsEngineSupport : JsEngineSupport() {
         val arguments = js("process.argv.slice(2).join(' ')") as String
         return arguments.split(' ').toTypedArray()
     }
+
+    override fun getMeasurer(): Measurer = NodeJsMeasurer()
+
+    override fun isSupported(): Boolean = isNodeJsEngine()
 }
 
 private fun hrTimeToNs(hrTime: dynamic): Long = (hrTime as Array<Double>).let { (seconds, nanos) ->
     seconds.toDuration(DurationUnit.SECONDS) + nanos.toDuration(DurationUnit.NANOSECONDS) }.inWholeNanoseconds
 
-internal inline fun nodeJsMeasureTime(block: () -> Unit): Long {
+private class NodeJsMeasurer : Measurer() {
     val process = js("process")
-    val start = process.hrtime()
-    block()
-    val end = process.hrtime()
-    return hrTimeToNs(end) - hrTimeToNs(start)
+    private var start: dynamic = 0.0
+    override fun measureStart() {
+        start = process.hrtime()
+    }
+
+    override fun measureFinish(): Long {
+        val end = process.hrtime()
+        return hrTimeToNs(end) - hrTimeToNs(start)
+    }
 }
+
+private fun isNodeJsEngine(): Boolean =
+    js("(typeof process !== 'undefined') && (process.release.name === 'node')") as Boolean
+
+internal actual var engineSupport: BenchmarkEngineSupport =
+    NodeJsEngineSupport