+++ runtime/jvmMain/src/kotlinx/benchmark/Utils.kt
@@ -15,5 +15,3 @@ internal actual fun String.readFile(): String {
 internal actual fun String.writeFile(text: String) {
     File(this).writeText(text)
 }
-
-internal actual inline fun measureNanoseconds(block: () -> Unit): Long = TODO("Not implemented for this platform")