+++ runtime/wasmJsMain/src/kotlinx/benchmark/Utils.kt
@@ -1,34 +1,31 @@
 package kotlinx.benchmark
 
-@JsFun("(d, precision, useGrouping) => d.toLocaleString('en-GB', { maximumFractionDigits: precision, minimumFractionDigits: precision, useGrouping: useGrouping } )")
-private external fun format(d: Double, precision: Int, useGrouping: Boolean): String
+private fun format(d: Double, precision: Int, useGrouping: Boolean): String =
+    js("d.toLocaleString('en-GB', { maximumFractionDigits: precision, minimumFractionDigits: precision, useGrouping: useGrouping } )")
 
 internal actual fun Double.format(precision: Int, useGrouping: Boolean): String = format(this, precision, useGrouping)
 
-internal actual fun String.writeFile(text: String): Unit = jsEngineSupport.writeFile(this, text)
+internal fun nodeJsEngineBinaryPath(): String = js("process.argv[0]")
 
-internal actual fun String.readFile(): String = jsEngineSupport.readFile(this)
+internal fun nodeJsEngineModulePath(): String = js("process.argv[1]")
 
-internal abstract class JsEngineSupport {
-    abstract fun writeFile(path: String, text: String)
+internal fun nodeJsGetDirName(filePath: String): String = path.dirname(filePath)
 
-    abstract fun readFile(path: String): String
-
-    abstract fun arguments(): Array<out String>
+internal fun throwModuleCannotBeImported(name: String) {
+    throw UnsupportedOperationException("Module $name cannot be imported in this environment")
 }
 
-@JsFun("() => typeof d8 !== 'undefined'")
-internal external fun isD8Impl(): Boolean
+/*
+Wasm JsFun expect something invokeable, so we have to return an arrow function.
+IIFE which returns a function that returns an input parameter (resolved module).
+It helps us to work around the issue that we cannot use await import in non-async arrow functions.
+(
+    (module) => {
+        return () => module
+    }
+)(await import("module"))
+ */
+internal const val LOAD_MODULE_PREFIX =
+    "((module) => () => module)(((typeof process !== 'undefined') && (process.release.name === 'node')) ? await import(/* webpackIgnore: true */'node:"
 
-internal val isD8: Boolean by lazy {
-    isD8Impl()
-}
-
-internal val jsEngineSupport: JsEngineSupport by lazy {
-    if (isD8) D8EngineSupport else NodeJsEngineSupport
-}
-
-internal external interface ExternalInterfaceType
-
-internal actual inline fun measureNanoseconds(block: () -> Unit): Long =
-    if (isD8) d8MeasureTime(block) else nodeJsMeasureTime(block)
+internal const val LOAD_MODULE_POSTFIX = "') : null)"