+++ runtime/wasmJsMain/src/kotlinx/benchmark/wasm/WasmBuiltInExecutor.kt
@@ -1,8 +1,6 @@
 package kotlinx.benchmark.wasm
 
 import kotlinx.benchmark.*
-import kotlinx.benchmark.internal.KotlinxBenchmarkRuntimeInternalApi
-
 
 private external interface JsAny
 
@@ -15,14 +13,15 @@ private fun id(p: Any): Any {
     return jsId(p as JsAny)
 }
 
-@KotlinxBenchmarkRuntimeInternalApi
-class WasmBuiltInExecutor(
-    name: String,
-    @Suppress("UNUSED_PARAMETER") dummy_args: Array<out String>
-) : CommonSuiteExecutor(name, jsEngineSupport.arguments()[0]) {
+/**
+ * Executes benchmarks in the built-in Wasm engine.
+ */
+internal class WasmBuiltInExecutor(name: String, configPath: String, xmlReporter: (() -> BenchmarkProgress)? = null) :
+    SuiteExecutor(name, configPath, xmlReporter),
+    RunAllBenchmarksExtension, CommonBenchmarkExtension {
 
     private val BenchmarkConfiguration.notUseJsBridge: Boolean
-        get() = "false".equals(advanced["jsUseBridge"], ignoreCase = true)
+    get() = "false".equals(advanced["jsUseBridge"], ignoreCase = true)
 
     @Suppress("UNCHECKED_CAST")
     private fun createJsMeasurerBridge(originalMeasurer: () -> Long): () -> Long =
@@ -37,4 +36,19 @@ class WasmBuiltInExecutor(
         val measurer = super.createIterationMeasurer(instance, benchmark, configuration, cycles)
         return if (configuration.notUseJsBridge) measurer else createJsMeasurerBridge(measurer)
     }
+
+    override fun runBenchmark(
+        benchmark: BenchmarkDescriptor<Any?>,
+        configuration: BenchmarkConfiguration,
+        parameters: Map<String, String>,
+        id: String,
+        progress: BenchmarkProgress
+    ): DoubleArray? = super.runBenchmark(benchmark, configuration, parameters, id, progress)
+
+    override fun run(
+        runnerConfiguration: RunnerConfiguration,
+        benchmarks: List<BenchmarkDescriptor<Any?>>,
+        start: () -> Unit,
+        complete: () -> Unit
+    ) = super<RunAllBenchmarksExtension>.run(runnerConfiguration, benchmarks, start, complete)
 }